Move Types.ts to server/osc-types.d.ts
