+++ packages/lang/src/internals/__tests__/fixtures/forward-stdout.txt
