+++ packages/lang/src/internals/__tests__/fixtures/io-class-syntax-error.txt
