+++ packages/lang/src/internals/__tests__/fixtures/io-compile-success.txt
