+++ packages/lang/src/internals/__tests__/fixtures/io-duplicate-class.txt
