+++ packages/lang/src/internals/__tests__/fixtures/io-extension-for-non-existent-class.txt
