+++ packages/lang/src/internals/__tests__/fixtures/io-programmatic-compile.txt
