+++ packages/lang/src/internals/__tests__/fixtures/script.scd
