+++ packages/lang/src/internals/__tests__/fixtures/trig-not-defined.txt
