+++ packages/server/src/__tests__/server-test.ts
@@ -1,7 +1,7 @@
 import _ from "lodash";
 import { EventEmitter } from "events";
 import Server from "../server";
-import { CallAndResponse } from "Types";
+import { CallAndResponse } from "../osc-types";
 
 describe("Server", function() {
   describe("default constructor", function() {
